assets and other issues

+++ src/pages/Users/UsersList.js
@@ -21,6 +21,7 @@ import TableLoader from "components/Common/TableLoader";
 import DeleteModal from "components/Common/DeleteModal";
 import UsersAddModal from "./UsersAddModal";
 import UsersEditModal from "./UsersEditModal";
+import { showErrorNotification } from "store/notifications/actions";
 function UsersList() {
 
   const [editModal, setEditUserModal] = useState(false);
@@ -179,7 +180,12 @@ function UsersList() {
 
   };
   const deleteUser = () => {
-    dispatch(deleteUsers(selectedUser._id));
+    const currentUser = JSON.parse(localStorage.getItem("authUser"));
+    if (selectedUser.email == currentUser.email){
+      dispatch(showErrorNotification("Cannot delete same user as you logged in right now!"));
+      setDeleteUserModal(false);
+    } else
+      dispatch(deleteUsers(selectedUser._id));
   };
 
   const statusUser = (user) => {
